SpeakFlow/pyproject.toml
[project]
name = "speakflow"
version = "0.1.0"
description = "Open-source voice input method — Speech to text with AI processing"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "openai>=1.60.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
    "aiofiles>=24.0.0",
    "Pillow>=11.0.0",
    "composio>=0.7.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[tool.ruff]
line-length = 120
